# db/userModel.js
const mongoose = require('mongoose')

const Schema = mongoose.Schema

const UserSchema = new Schema({
  email: { type: String, required: true, unique: true },
  password: { type: String, required: true }
})

module.exports.UserModel = mongoose.model('User', UserSchema)